+++ platformio/managers/core.py
@@ -27,7 +27,7 @@ CORE_PACKAGES = {
     "contrib-piohome": ">=3.1.0-rc.2,<3.2.0",
     "contrib-pysite": "~2.%d%d.0" % (sys.version_info[0], sys.version_info[1]),
     "tool-pioplus": "^2.6.1",
-    "tool-unity": "~1.20403.0",
+    "tool-unity": "~1.20500.0",
     "tool-scons": "~2.20501.7" if PY2 else "~3.30102.0",
     "tool-cppcheck": "~1.189.0",
     "tool-clangtidy": "^1.80000.0",